requirements.txt
accelerate==1.1.1
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
diffusers==0.31.0
einops==0.8.0
filelock==3.13.1
fsspec==2024.10.0
gmpy2==2.1.2
huggingface-hub==0.26.2
idna==3.7
importlib_metadata==8.5.0
Jinja2==3.1.4
kornia==0.7.4
kornia_rs==0.1.7
MarkupSafe==2.1.3
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pillow==11.0.0
pip==24.2
psutil==6.1.0
PySocks==1.7.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.1.0
sympy==1.13.2
timm==1.0.11
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.67.0
triton==2.1.0
typing_extensions==4.11.0
urllib3==2.2.3
wheel==0.44.0
zipp==3.21.0
